Add files via upload
Just scan this QR code and you'll be directed to a website named MyWebAr.com. So give access of your phone camera and keep your camera to the QR code and you'll be mesmerize to see the India map in 3D view, which actually is not there in the reality but because of MyWebAr.com you can see it.